frontend: extract types from BaseHorseEventForm
